requirements.txt
pip-chill==1.0.1
python-dotenv==0.20.0
sentinelsat==1.1.1
